Clean up unnecessary abstraction
