--- support/cas-server-support-edit-attributes-core/src/main/java/org/apereo/cas/attributes/AbstractEditableAttributeRepository.java
@@ -1,11 +1,0 @@
-package org.apereo.cas.attributes;
-
-/**
- * @author Marcus Watkins
- * @since 5.2
- *
- */
-public abstract class AbstractEditableAttributeRepository implements EditableAttributeRepository {
-
-
-}
